[MAIN] Note upload and download - server side. (#552)
* Note upload and download - server-side.

Download:
Fetching data from the server and encoding it through 3 layers. After that, save notes to CoreDatawhich are connected to the specified session.
Upload:
Fetching data from Core Data and coding it through 3 layers. After that, send it to the server inside the session struct.

EditNoteViewModel was changed, that it can propagate local changes on the server - using self.sessionUpdateService.updateSession(session: session). SingleSessionSynchronizer was changed, to handle the notes inside the session.