Swift lint and fastlane (#564)
* SwiftLint was added to project and new lane which uses SwiftLint.

Every Date is now handled by the DataBuilder. Whenever user tries to use Date() he will get an error.
FastLane has a new lane, called code_format which aims to check code.
SwiftLint also was provided with ability to detect prints, which we don't want to have in the production.