Logic for requesting location changed.
Now when session if fixed and indoor the location screen is skipped. When sesssion is fixed and outdoor, location screen is shown as a reminder and an option for the user. The only requirement is to allow location when doing mobile session without disable mapping. In this commit, we are handling a special case where user could have location services turned off at all and never get Apple popup about location.