Condition for graph update changed along with threshold for simplyfying data points (#979)
Currently, after passing 24 hours of collecting measurements, we were always having the same number of total measurements. The reason for this is that when we collect a new measurements, we always delete the last ones. This makes the total count always the same.
Next thing was about time correctness on the graph and session itself. There was some discrapency, becouse of our averaging service and the fact that the time is taken from x point of latest point (which was previously averaged). To prevent such scenario we are increasing threshold to 1500 as the max value for measurments should not exceed 1440 for fixed session.